SPLICE-730 Cleaning up BasicDatabase operations, added a few java docs, code cleanup
