+++ splice_machine/src/main/java/com/splicemachine/derby/impl/job/fk/FkJobSubmitter.java
@@ -29,7 +29,6 @@ import com.splicemachine.protobuf.ProtoUtil;
 import com.splicemachine.si.api.txn.TxnView;
 import org.apache.log4j.Logger;
 import org.sparkproject.guava.collect.ImmutableList;
-
 import java.util.List;
 
 /**
