+++ splice_machine/src/main/java/com/splicemachine/derby/impl/sql/compile/BroadcastJoinStrategy.java
@@ -95,7 +95,7 @@ public class BroadcastJoinStrategy extends HashableJoinStrategy {
         if(td==null) return false;
 
         /* Currently BroadcastJoin does not work with a right side IndexRowToBaseRowOperation */
-        if(JoinStrategyUtil.isNonCoveringIndex(innerTable)) {
+        if(isNonCoveringIndex(innerTable)) {
             return false;
         }
