--- splice_machine/src/main/java/com/splicemachine/derby/impl/sql/compile/JoinStrategyUtil.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2012 - 2016 Splice Machine, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use
- * this file except in compliance with the License. You may obtain a copy of the
- * License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed
- * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package com.splicemachine.derby.impl.sql.compile;
-
-import com.splicemachine.db.iapi.sql.compile.AccessPath;
-import com.splicemachine.db.iapi.sql.compile.Optimizable;
-import com.splicemachine.db.iapi.sql.dictionary.ConglomerateDescriptor;
-
-public class JoinStrategyUtil {
-
-    public static boolean isNonCoveringIndex(Optimizable innerTable) {
-        try {
-            AccessPath path = innerTable.getCurrentAccessPath();
-            if (path != null) {
-                ConglomerateDescriptor cd = path.getConglomerateDescriptor();
-                if (cd != null && cd.isIndex() && !innerTable.isCoveringIndex(cd)) {
-                    return true;
-                }
-            }
-        } catch (Exception e) {
-            throw new IllegalStateException("could not determine if index is covering", e);
-        }
-        return false;
-    }
-
-}
