+++ splice_machine/src/main/java/com/splicemachine/derby/stream/control/ControlDataSetProcessor.java
@@ -59,7 +59,6 @@ import com.splicemachine.storage.Partition;
 /**
  * Local control side DataSetProcessor.
  *
- * @author jleach
  */
 public class ControlDataSetProcessor implements DataSetProcessor{
     private long badRecordThreshold=-1;
